Extract EULA scroll progress handling
